camera-legacy: ICP-HFI: Raise maximum power collapse poll retries
It is observed on SDM845 that the HFI may reach the idle (wfi)
state after a lot of time because, sometimes, hardware tear-down
may be slow due to error conditions ongoing while BW is low:
in this case the Spectra HFI would not call WFI idle in time.

If we tear it down and reset it while not idling, we may get a
bad status on other hardware governed by the HFI and a slow
down for recovery - or a bad crash - the next time we try to
open the camera hardware from the userspace.

For this reason, raise the max poll trials, so that we give
more (and hopefully enough) time to the HFI to finish the hw
tear-down process before resetting the HFI's CPU, leaving the
entire thing in a clean state and ready for the next operation.